refactor[/aurhors] продолжить Authors
Продолжена работа над AuthorForm
Исправлена опечатка в  названии файла Authoform.tsx
Доработаны данные в authors.json
