+++ backend/src/data/authors.json
@@ -1,14 +1,26 @@
 [
     {
         "id": "001",
-        "name": "Толстой Лев Николаевич"
+        "surname": "Толстой",
+        "name": "Лев",
+        "secondName": "Николаевич",
+        "fullname": "Лев Николаевич Толстой",
+        "abbreviatedName": "Л. Н. Толстой"
     },
     {
         "id": "002",
-        "name": "Пушкин Александр Сергеевич"
+        "surname": "Пушкин",
+        "name": "Александр",
+        "secondName": "Сергеевич",
+        "fullname": "Пушкин Александр Сергеевич",
+        "abbreviatedName": "А. С. Пушкин"
     },
     {
         "id": "003",
-        "name": "Чехов Антон Павлович"
+        "surname": "Чехов",
+        "name": "Антон",
+        "secondName": "Павлович",
+        "fullname": "Антон Павлович Чехов",
+        "abbreviatedName": "А. П. Чехов"
     }
 ]
