+++ frontend/src/features/authors/ui/AuthorForm/AuthorForm.scss
@@ -32,3 +32,7 @@
 .author-form__input--auto-generated {
     cursor: no-drop;
 }
+
+.form__label__error {
+    color: var(--red);
+}
